requirements.txt
python-telegram-bot==20.3
openai==0.27.0
apscheduler==3.8.1  # Adicionando a versão específica, se necessário
nest-asyncio==1.5.4  # Adicionando a versão específica, se necessário
pytz